added ant movement & ant rotation

+++ Coordinates.h
@@ -1,9 +1,9 @@
 #ifndef COORDINATES_H
 #define COORDINATES_H
 struct Coord{
-    int x=10, y=10, id = 0;
+    float x=10, y=10, id = 0;
     Coord(){};
-    Coord(int x, int y, int id) : x(x), y(y), id(id) {};
+    Coord(float x, float y, float id) : x(x), y(y), id(id) {};
     Coord(const Coord& coord) : x(coord.x), y(coord.y), id(coord.id) {};
     ~Coord(){};
     Coord& operator =(const Coord& coord)
